Split C codegen type mapping tests
