+++ docs/COMPLETION_AUDIT.md
@@ -2968,6 +2968,10 @@ and do not assume Phase 4 is ready without evidence.
   `src/lexer/tests/syntax_examples.rs`, preserving Zen function/import,
   declaration, UFC, pattern-match, and method-call tokenization examples while
   keeping core token/operator coverage focused.
+- C codegen type mapping tests now live in
+  `src/codegen/c/tests/type_mapping.rs`, preserving primitive, string, pointer,
+  named, and function-pointer type mapping coverage while keeping program
+  generation and helper emission tests focused.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
