+++ docs/PHASE_PLAN.md
@@ -2703,6 +2703,10 @@ checked-in docs, tests, and commits only.
   `src/lexer/tests/syntax_examples.rs`, keeping Zen function/import,
   declaration, UFC, pattern-match, and method-call tokenization examples
   separate from core token/operator coverage.
+- C codegen type mapping tests now live in
+  `src/codegen/c/tests/type_mapping.rs`, keeping primitive, string, pointer,
+  named, and function-pointer type mapping coverage separate from program
+  generation and helper emission tests.
 
 ## Current Phase
